pyproject.toml
[tool.poetry]
name = "gendiff-by-JhonnyT"
version = "0.3.1"
description = ""
authors = ["Evgenii T"]
packages = [
  { include = "gendiff" },
]


[tool.poetry.dependencies]
python = "^3.8"
PyYaml = "^5.3.1"


[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
pytest = "^6.1.2"
pytest-cov = "^2.10.1"
coverage = "^5.3"


[tool.poetry.scripts]
gendiff =  'gendiff.scripts.gendiff:main'


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
